test(e2e): setup for ci

+++ packages/testing/src/playwright/playwright-launcher.ts
@@ -2,20 +2,28 @@ import { Config } from '@jest/types';
 import { setup, teardown } from 'jest-process-manager';
 import onExit from 'signal-exit';
 
-import { server } from './e2e-server.config';
+import { server } from './playwright-server.config';
 
 const { globalSetup, globalTeardown } = require('jest-playwright-preset');
 
 let serverSetupPromise: Promise<void> | undefined;
 
 export const destroyServer = async (globalConfig?: Config.GlobalConfig) => {
+  if (server.environment !== 'playwright') {
+    return;
+  }
+
   serverSetupPromise = undefined;
   await teardown();
   await globalTeardown(globalConfig);
 };
 
 export const setupServer = async (globalConfig: Config.GlobalConfig) => {
-  await setup([server.server]);
+  if (server.environment !== 'playwright') {
+    return;
+  }
+
+  await setup([server.config]);
 
   onExit(() => {
     destroyServer().then(() => {
